libwasi: move path api functions into a separate file
