+++ libwasi/CMakeLists.txt
@@ -14,6 +14,7 @@ set(lib_wasi_sources
 	"wasi_clock.c"
 	"wasi_fd.c"
 	"wasi_environ.c"
+	"wasi_path.c"
 	"wasi_poll.c"
 	"wasi_proc.c"
 	"wasi_random.c"
